Move Marshaller to event package
